add NYMEA-COMMERCIAL license file
The license file is an important piece of information for whoever is
building packages.

When Yocto builds a package, it adds all the license files to
$DEPLOY_DIR/licenses/<package>/generic_<license_name>.

They could also be technically added to a rootfs.

Ideally, this should be a plain text file and not a PDF but I don't have
the sources of the PDF and can't be bothered to manually do it
(especially since pdftotext does not like the two columns layout).

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>